requirements-dev.txt
-e .
black
build
deep-trainer
mypy
pylint
torchvision
twine
